requirements.txt.txt
onnxruntime-gpu>=1.18.0
opencv-python>=4.9.0
pyrealsense2
numpy>=1.23
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
streamlit
PyYAML
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch
pillow
pandas
openpyxl>=3.1.0
xlrd>=2.0.1
matplotlib
scikit-learn
tqdm
schedule
python-dateutil
seaborn
tabulate
ultralytics>=8.0.0
plotly
